+++ tornadowebapi/deserializers/basic_rest_deserializer.py
@@ -1,6 +1,6 @@
 from tornadowebapi.schema import Schema
 from tornadowebapi.schema_fragment import SchemaFragment
-from tornadowebapi.singleton_resource import SingletonResource
+from tornadowebapi.singleton_schema import SingletonSchema
 from tornadowebapi.traitlets import OneOf
 from .base_deserializer import BaseDeserializer
 
@@ -16,7 +16,7 @@ class BasicRESTDeserializer(BaseDeserializer):
 
         if issubclass(resource_class, Schema):
             resource = resource_class(identifier=identifier)
-        elif issubclass(resource_class, (SchemaFragment, SingletonResource)):
+        elif issubclass(resource_class, (SchemaFragment, SingletonSchema)):
             resource = resource_class()
         else:
             raise TypeError(
